fix: :bug: ISSUE-7 WorkflowRun status error
The previous configuration for  WorkflowRun.kt -> RunStatus enum class
only allowed for 3 statuses:
- QUEUED
- IN_PROGRESS
- COMPLETED

There are more status states as defined by the github actions/runs api
https://docs.github.com/en/rest/reference/actions#list-workflow-runs-for-a-repository

"GitHub can set a status of waiting or requested."

In the case where the error was detected the status of `waiting` was
returned. This lead to the error:

"Error: Array contains no element matching the predicate."

By adding the enum property `WAITING("waiting")` the predicate can be
matched.
As a pre-emptive measure I'm also adding `requested`.

ISSUE-7